+++ dev/towerdefense/global/default.js
@@ -3,7 +3,7 @@ Imported.UFCTowerDefense = true;
 
 var UFC = UFC || {};
 UFC.UFCTD = UFC.UFCTD || {};
-UFC.UFCTD.VERSION = 1.0;
+UFC.UFCTD.VERSION = 1.2;
 UFC.UFCTD.ALIAS = UFC.UFCTD.ALIAS || {};
 
 var $dataTDEnemy = $dataTDEnemy || {};